merge: dev to main

+++ database/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Boolean, DateTime, ForeignKey
+from sqlalchemy import Column, Integer, String, Boolean, DateTime, ForeignKey, BigInteger
 from sqlalchemy.orm import relationship, declarative_base
 from datetime import timezone, datetime
 
@@ -8,7 +8,7 @@ Base = declarative_base()
 class User(Base):
     __tablename__ = "users"
 
-    user_id = Column(Integer, primary_key=True)
+    user_id = Column(BigInteger, primary_key=True)
     language = Column(String, default="en")
     created_at = Column(DateTime(timezone=True),
                         default=lambda: datetime.now(timezone.utc), nullable=False)
@@ -25,7 +25,7 @@ class Subscription(Base):
     __tablename__ = "subscriptions"
 
     id = Column(Integer, primary_key=True, autoincrement=True)
-    user_id = Column(Integer, ForeignKey("users.user_id"), nullable=False, unique=True)
+    user_id = Column(BigInteger, ForeignKey("users.user_id"), nullable=False, unique=True)
     is_premium = Column(Boolean, default=False)
     premium_start = Column(DateTime, nullable=True)
     premium_end = Column(DateTime, nullable=True)
